requirements.txt
discord.py==2.7.0
yt_dlp==2026.2.21
